+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Refactored `RedisTask` to be more flexible
   - Added Redis helper methods for (un)formatting `HSET` and `HGET` style commands
   - Rewrote documentation for this task
+- `HarvestUpdateTask` will now make sure the `Harvest.UniqueIdentifier` key is uniquely indexed on the target and `metadata` collections
 
 ## 0.6.3
 - Added the Environment package which provides a way to store configuration information made available to tasks